Split playground sync helpers
